Float32

+++ models.py
@@ -112,8 +112,8 @@ class GeoCLIP(nn.Module):
         gps_location_similarity = location @ location.t()
         
         # Normalize
-        gps_features_similarity = (gps_features_similarity / gps_features_similarity.norm(dim=1, keepdim=True)).to(torch.float64)
-        gps_location_similarity = (gps_location_similarity / gps_location_similarity.norm(dim=1, keepdim=True)).to(torch.float64)
+        gps_features_similarity = (gps_features_similarity / gps_features_similarity.norm(dim=1, keepdim=True)).to(torch.float32)
+        gps_location_similarity = (gps_location_similarity / gps_location_similarity.norm(dim=1, keepdim=True)).to(torch.float32)
         
         gps_sim = gps_features_similarity @ gps_location_similarity.t()
         gps_sim = (gps_sim + 1) / 2
